shill: portal detector: refactor retry logic
This patch refactors the management of portal detection retries such
that the computation of retry intervals and delays and their associated
state are now completely internal to PortalDetector.

Device is still in charge of triggering the retries based on the
connection state of the service. Future patches will migrate this logic
outside of Device or encapsulate it into a connection state machine.

This patch also changes the types used in favor of base::Time and
base::TimeDelta.

BUG=b:182745554
BUG=b:184036481
TEST=Unit tests. Checked network validation on WiFi and ethernet.
Checked portal detection on coffee shop WiFi. Checked probe retrial on
WiFi hotspot with no Internet.

Change-Id: I9f6051ac658bd71304b693a0dbb3c28a00c84910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2842293
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>