shill: Cellular: Disassociate Service when Capability is destroyed
When the MM1.Modem object is destroyed, the associated Capability
instance is also destroyed. At this point no Service should be
associated with the Cellular object.

By not clearing Cellular.service_, when a new Capability was
created, the new activation state was being incorrectly set on
the old Service instance. This may also be true for the Operator
(and therefore the Name).

BUG=b:185513569,b:185623571
TEST=unit tests, tast tests, manual test

Change-Id: I7270ad035447fb3628903f231983c781c57e51b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2844349
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Steven Bennetts <stevenjb@chromium.org>