arc: libvda: decode: use Rc to reference VDA connection from decoder sessions
The use of a direct reference to the VDA instance in the decoder session
severely limits the way a decoder can be used without using unsafe code:
in order to guarantee lifetimes, the VDA instance must be explicitly
allocated in such a way that the borrow checker can see that it lives
longer than any object using it (i.e. it cannot be owned by an object).

This is going to be a limiting factor for crosvm as we want to use
dynamic decoder backends, which may or may not use a VDA instance,
making it impossible to not reference libvda in code that ought to be
generic without getting an error from the borrow checker. Replace the
references with a Rc to the VdaConnection: this guarantees at runtime
that the connection will be living for as long as any session is open,
and removes the need to use lifetimes with decoder sessions, all for for
an imperceptible runtime cost.

BUG=b:169295147
TEST=Video plays correctly on hatch-arc-r.

Cq-Depend: chromium:2845633
Change-Id: I9e0eae766602e8c2471d92433c278666677f21ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2838714
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>