init: setting tracing_on to 0 after tracefs initialization
/sys/kernel/debug/tracing/tracing_on is 1 right after the kernel
initializes tracefs, and it will be reset to 0 by Android's init running
in the ARC++ container. This inconsistency leads to confusion that
tracing_on can be 0 or 1 on different devices. We explicitly setting it
to a reasonable initial value of 0 after tracefs is initialized by
upstart.

BUG=b:182601127
TEST=check /sys/kernel/debug/tracing/tracing_on on a DUT without ARC++.

Change-Id: I98754669c1c1585107c857a4fdcd142b99d7c199
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2846552
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>