patchpanel: Allow VPN traffic from namespace to local source
ConnectNamespace API currently does not support traffic from
guests when VPN is expected. This happens because
ConnectNamespace API provides only a single interface for
both traffic to the physical network and traffic to the local
source.

However, it only has a single “-i” rule which is ambiguous
and is unable to distinguish between the cases just mentioned.
Traffic going back to the local source should not be routed
through VPN.

This patch adds a rule to exclude return traffic from a
connected namespace to be routed through a VPN.

BUG=b:171363265
TEST=emerge-rammus-arc-r patchpanel
TEST=iptables -t mangle -S;
TEST=units;
TEST=fuzzer;

Change-Id: I4e68fb7a571c215ef7642ef787e559bbd3477684
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2847522
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>