vm_tools: concierge: Implement ReclaimVmMemory
This CL allows the client e.g. Chrome to ask Concierge to reclaim VM
memory. In the handler, Concierge writes "shmem" to /proc/pid/reclaim
to move the VM's memory to zram to free up memory. We already do that
for Chrome renderer and ARC container processes (to reduce Chrome UI
jank when ARC is not actively in use, for example), and this is ARCVM
port of the feature.

BUG=b:186278424
TEST=concierge_client

Change-Id: I24b31963320f34db15e429210bd6de764f553713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2851841
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>