dns-proxy: Start default proxy unconditionally
When DNS proxy is enabled, USER DNS traffic need to
be redirected properly to a DNS proxy. If VPN is off,
the traffic can be redirected to the system proxy or
the default proxy. However, if VPN is on, the traffic
must be redirected to the default proxy as VPN is not
applied to traffic of the system proxy.

In order to have cleaner DNS redirection rules, always
have the default proxy run. This allows the rules to
always redirect USER DNS traffic to the default proxy
as opposed to redirecting to the system proxy when the
default proxy is not running.

BUG=b:171363265
TEST=emerge-rammus-arc-r dns-proxy;
TEST=dig <VPN_ONLY_HOSTNAME>; dig <HOSTNAME>;
     on crostini, chronos, cups, root.
TEST=Toggle VPN on and off.
TEST=Turn crostini on and off.

Change-Id: I6d4226e3d10775ad6b9d6a6dc561a23864de4285
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2853785
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>