libec: Increase FpContextCommand polling timeout to 7 seconds
Sensor initialization (especially when finger is kept on it) can take
up to 5 seconds. Currently, we are waiting 2 seconds for the result of
asynchronous command. If it takes more than 2 seconds to finish command
following message is printed in biod.LATEST:

[ec_command_async.h(76)] Timed out polling for command 0x406

This patch increases polling timeout to 7 seconds. This value is enough
to get rid of error when starting biod with finger on sensor. This patch
also extends FpContextCommand_v1 unit test (and performs necessary code
changes) to make sure that FpContextCommand uses appropriate polling
timeout.

BUG=b:185244590
TEST=Stop biod using 'stop biod'. Issue 'start biod' while keeping
     finger on sensor. Make sure that biod.LATEST log doesn't contain
     message about polling timeout for 0x406 command.
TEST=FEATURES="test" emerge-volteer libec

Change-Id: I01897125e8ecfe68209166abe23489fc8b3cea6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2850158
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>