shill: Replace rmnet_ipa0 with rmnet_data0
Now that multiplex is supported in ModemManager, the net interface for
QRTR based modems is only created when the bearer is created. Shill
doesn't know how to handle this, since it assumes that the modem will
send the net interface information as soon as the modem is created, and
not later.
ModemManager sends the name of the parent interface(rmnet_ipa0) when the
modem is created, and sends a signal with the new net link(rmnet_data0)
when the bearer is created. Since shill only needs the link name, this
little hack tells shill to use the link name instead of the parent name.

BUG=b:175305412
TEST=test_that $LAZOR_IP:$LAZOR_PORT suite:cellular_ota --fast
TEST=Verify internet connection in lazor when connecting and
disconnecting multiple times. Disable and enable Mobile data without
issues.

Cq-Depend: chromium:2855769
Change-Id: I246af553fe6b472bffa0ec3d4f670b8ce4937dff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2855665
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>