power: Read display SoC from EC
Currently, the power manager calculates the display SoC from the
remaining charge, the full charge, the shutdown SoC, and the full
factor.

To get a consistent view of the display SoC in all power states,
the EC will calculate the display SoC and export it. This patch
will make the power manager read the display SoC, the full factor,
and the low battery percentage from the EC.

If the EC is found not exporting the display SoC, the power manager
falls back to the current behavior.

BUG=b:174433637, b:181506409, b:80270446, b:109954565
TEST=On Storo & CoachZ, verified display SoCs match between EC and
power_supply_info.
TEST=On Storo, verify power_supply_info prints 'status Fully charged'
when display SoC reaches 100.
TEST=On Storo, verify system shuts down when display SoC reaches 0.
TEST=On Storo, verify system behaves expectedly if EC doesn't support
CONFIG_BATTERY_EXPORT_DISPLAY_SOC.

Change-Id: I5bd1371f2569d21d55df1b50a3d709b98bbf0325
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2853269
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>