rmad: Add IsRepeatable() attribute in state handlers
IsRepeatable() decides whether we can go back to a previous state, or
abort the RMA process. Going back to the previous state is allowed only
when the current state and previous state are both repeatable. Aborting
is allowed only when we can go back to the welcome screen, which implies
that we're not allowed to abort RMA when we enter the first unrepeatable
state.

BUG=b:182242749
TEST=emerge-volteer rmad;
     cros_run_unit_test --board volteer --packages rmad

Change-Id: I21e1fb9fefc4678d95718f8e24d379b8a4217d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2853782
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>