vm_tools: concierge: add CPU cluster and capacity support for Crostini
On systems with heterogeneous cores (e.g. Arm's big.LITTLE), organize
the vCPU threads into clusters based on the host kernel's topology
physical_package_id property.

Also add the crosvm --cpu-capacity arg when the host CPU sysfs entries
expose the cpu_capacity field.  This will result in the addition of the
capacity-dmips-mhz value in the devicetree cpu nodes, allowing the guest
kernel to make better scheduling decisions on platforms with
heterogeneous cores (big.LITTLE).

For example, on a trogdor board, there are 6 little cores (0-5) with a
capacity of 452, and 2 big cores (6-7) with a capacity of 1024, which
results in a crosvm command line argument of:

  --cpu-capacity 0-452,1=452,2=452,3=452,4=452,5=452,6=1024,7=1024

On non-arm systems, these options will be accepted but ignored (support
may be added later for reporting via ACPI if needed).

BUG=b:182198842
TEST=Run Crostini; verify phyiscal_package_id in guest
TEST=lstopo in guest on kevin shows two clusters
TEST=cat /sys/devices/system/cpu/cpu*/cpu_capacity
TEST=dtc -I fs -O dts /sys/firmware/devicetree/base

Cq-Depend: chromium:2762298
Change-Id: I70337f98f7e4afac6fa036644f0b56e1b48ec575
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2857362
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>