crash: Change lacros parsing order.
When a JavaScript error is generated by crash_reported, it attaches a
set of logs to the report. For Lacros browser, we need to attach a
different set of logs. This change sets the log collection priorities to
pick the right set of logs.

BUG=b:186637027
TEST=Generated JavaScript error in Lacros, ensured correct logs

Change-Id: I0754b75e5b2db4cbfa03c5781d6f7fc80f774f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2859715
Tested-by: Gabriel Szlechtman <gabrielsz@google.com>
Commit-Queue: Gabriel Szlechtman <gabrielsz@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>