power: Disable CrOS EC device event on suspend entry
Currently, PCHG can send a device event any time. When a signal is
sent to the host while it's attempting to suspend, suspend is delayed
by 10 seconds.

A solution is to make Powerd disable a corresponding PCHG event (i.e.
EC_DEVICE_EVENT_WLC) when it's announcing an upcoming suspend.

When a suspend is cancelled device event for WLC is re-enabled and
suspender prints:

[suspender.cc(495)] Starting request 79495169
[suspender.cc(486)] Aborting request in response to event UserActivity
[suspender.cc(538)] Finishing request 79495169 unsuccessfully after 0s

BUG=b:182973695, b:173235954
TEST=Verify CoachZ suspends without a delay by closing lid, keyboard,
idle, and powerd_dbus_suspend.
TEST=Device event for WLC is re-enabled when suspend is cancelled.

Change-Id: I41d1b2762de2a0162c1ce3122fc80cd48183bfee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2862160
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>