metrics: memd: allow fstatat64
The combination of glibc 2.32 and Rust 1.51 cause this code to start
emitting calls to __fxstatat, which thus requires the addition of
fstatat64 here.

newfstatat was added on the x86 side here:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2849853.

CQ testing didn't seem to catch the lack of this entry in the allowlist,
however.

BUG=chromium:1192744, b:186803271
TEST=CQ

Change-Id: Ic50b226b7331afaf73bbab4feec887c42770c1f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2863311
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>