os_install_service: move dbus constants to system_api
Move the dbus string constants from os_install_service to a new file
under system_api. This will allow the constants to be used in client
code in the browser.

To avoid duplicating the constants between the Rust and C++ code, add
a build.rs file to os_install_service that parses the C++ header and
converts the constants to Rust. This is based on similar code in
vm_tools/crostini_client/build.rs.

BUG=b:182386612
TEST=FEATURES=test emerge-$board chromeos-base/system_api os_install_service

Cq-Depend: chromium:2869148
Change-Id: I8003e5a4d8c5256fac356b5fdff7b05e3a38b1dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2869089
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>