mount-encrypted: Fix resize units while spawning the filesystem resizer
In CL:2590694, instead of formatting a minimal filesystem and
resizing it, we combine the two steps into a single ext4 format call.
This renders the subsequent SpawnResizer() call to be redundant.

But the CL also added a logical bug in the number of blocks the fs
resizer is called with (bytes instead of the block size) and immediately
fails. The impact of this bug is that devices with undersized stateful
partitions (b:172344853) will not be resized accurately since resize2fs
is called with an absurdly large number.

BUG=b:186636538
TEST=platform_MountEncrypted, we no longer see the resize2fs abort in
     dmesg: resize2fs reports:
     "The filesystem is already 769172 (4k) blocks long.  Nothing to
     do!"

Change-Id: If123e91831bf741cc452de9c61d47689c919ff17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2872427
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>