shill: Check selected_service exists on neighbor event
In WiFi::OnNeighborReachabilityEvent(), checks if the device has a
selected service before proceeding. If not, we don't need to do anything
on a link event.

Also fixes some style issues that should be done in crrev.com/c/2821075
but was reintroduced by a bad rebase.

BUG=b:147256664
TEST=unit_tests

Change-Id: Ia6abd4ec5a1342d1732123582984a771acac3482
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2879796
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>