ml: add "socket: arg0 == AF_UNIX" to the WebPlatform hwr policy.
On boards that does not support hwr, loading the model will lead to
socket syscall. This causes ml-service to crash. Now we fix this in
this CL by adding socket to the policies.

Note that on hwr-supported boards, there is no issue like this.

BUG=chromium:1206207
TEST=on pyro, it now can correctly reject the promise of creating new
TEST=hwr models. On nocturne and strongbad, the API still works.

Change-Id: I70d7b10d7de7193c89b92251c8356fdb49d19e44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2875058
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>