patchpanel: Skip DNS when parsing traffic counters
When a VPN is just connected or disconnected, there
is a very small timing (<1s) when DNS is not working.
It is possible that traffic counter is querying
iptables while DNS is not working.

This can delay the creation of DNS proxy redirection
rules, extending the time of DNS not working.

Stop querying DNS when parsing iptables counters by
passing "-n" option.

BUG=None
TEST=Toggle VPN on off with DNS proxy.
TEST=units

Change-Id: Ic44ac0ffc1f015e0f3040a7f2adb0f6719161d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2886280
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>