metrics: use int64_t instead of int for structured metrics
Some clients need ints larger than 32 bits on all platforms, so this
changes int metrics to be recorded as int64_t instead of int.

This should be a backwards compatible change with existing recorded
metrics, and is already supported in the proto used for upload.

BUG=chromium:1148168
TEST=compiles, codegen uses int64_t in all palces.
     unit tests being added in chromium

Change-Id: I2c9efd9761d15653fa543dbc244930dfa341552c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2886291
Reviewed-by: Yu Liu <yudiliu@google.com>
Commit-Queue: Yu Liu <yudiliu@google.com>
Tested-by: Yu Liu <yudiliu@google.com>