arc: libvda: encode: use Rc to reference VEA connection from decoder sessions
The use of a direct reference to the VEA instance in the encoder session
severely limits the way an encoder can be used without using unsafe
code: in order to guarantee lifetimes, the VEA instance must be
explicitly allocated in such a way that the borrow checker can see that
it lives longer than any object using it (i.e. it cannot be owned by an
object).

This is going to be a limiting factor for crosvm as we want to use
dynamic encoder backends, which may or may not use a VEA instance,
making it impossible to not reference libvda in code that ought to be
generic without getting an error from the borrow checker. Replace the
references with a Rc to the VeaConnection: this guarantees at runtime
that the connection will be living for as long as any session is open,
and removes the need to use lifetimes with encoder sessions, all for for
an imperceptible runtime cost.

BUG=b:169295147
TEST=arc.VideoEncodeAccel.h264_192p_i420_vm passes on hatch-arc-r

Cq-Depend: chromium:2887525
Change-Id: I4e20b02d4cec7fddea96f337c2d48ce0bebdb109
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2887389
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>