shill: vpn: Check Manager::IsOnline() when creating VPNService
The current code assumes that manager is online when creating a
VPNService, and sets |last_default_physical_service_online_| to true.
So if there is no online service when the VPNService is created and
there comes one after that, VPNService will send out a default service
changed event to the driver instead of an up event. This patch fixes
this issue.

BUG=None
TEST=unit_tests

Change-Id: If3dcb4d596148d0091d08fb9edbf754de81e5684
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2890151
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>