login: Implement LoadShillProfile
LoadShillProfile will be used by chrome to signal that the shill profile
for a user should be loaded.
A follow-up CL will decouple loading the shill profile from the start of
the user session and will couple it to the new impulse
load-shill-profile instead.

BUG=b:183084821
TEST=cros_run_unit_tests --board ${BOARD} --packages chromeos-login

Change-Id: Ica0094fb9f66b20eaf0904c3eb7a00b1f4a1ca9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2891562
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Tested-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>