shill: Load the user profile on load-shill-profile event
Change the event that triggers loading of the shill user profile from
'start-user-session' to 'load-shill-profile'.
Before this change, the shill profile was loaded implicitly when the
Chrome OS user session starts.
After this change, the shill profile is loaded when explicitly triggered
by chrome through session_manager.
This allows chrome to ensure that a user session password has been saved
in case it is necessary for network policies.

BUG=b:183084821
TEST=manual, deployed the CL chain and Chrome CL CL:2891583 and \
checked that shill profile is still loaded on user sign-in. Checked
that $BUG is resolved with real PEAP-MSCHAPv2 setup.

Change-Id: I23634fde61cb53312b3030f5873c14975323d222
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2891563
Tested-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>