biod: Temporarily disable maintenance timer
The maintenance timer could be triggered during an auth/unlock session.
This has been shown to silently abort the auth/unlock session, which
manifests to the user as fingerprint unlock just not working.

We need to disable this timer until we can find a safe way for
these operations to coexist.

BUG=b:184783529,b:187951992
TEST=FEATURES=test emerge-hatch biod

Change-Id: I293faa7b833f8726e2699916a08c1f1efc3b0621
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2892333
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>