shill: vpn: wg: Try kernel interface at first
This patch adds the logic that, when connecting to a wireguard vpn
service, the WireguardDriver will try to create a wg interface in the
kernel at first. If that fails, then the driver will spawn the userspace
wg program.

Note that to configure a kernel wg interface, wireguard-tools needs to
be ran with CAP_NET_ADMIN.

BUG=b:177876632
TEST=Built the wg kernel module for octopus, manually verified that the
  connecting behavior before and after insert that kernel module.

Change-Id: I99dbbca20f1466cb02da941e4faae89e2b6b8420
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2894071
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>