ml: update text suggester enums
After rolling the updates to text_suggester.mojom to chromium, there
were a few nits that needed to be addressed, see crrev/c/2890015 for
more details.

This CL fixes the nits outlined in the CL above. Once this CL has landed
the updates will be rolled to the CL above.

BUG=chromium:1146266
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-base/ml

Change-Id: If9746841387dafe7343c914a3291e493177a37ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2896664
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>