arc: Move arc-sensor script to arc-setup
nsenter usage in arc-sensor.conf is unsafe as it doesn't run in an
Android SELinux context.
When Android init is compromised, it can bind-mount an executable on
/system/bin/touch and run it in a Chrome OS SELinux context.

To remove it, move the script to arc-setup.

This also makes the execution time faster from 300ms to 2ms.

BUG=chromium:1163124
TEST=App can read sensor data

Change-Id: I7a2aad6de001aaab493090c6e59337b400a13554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2896440
Tested-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>