shill: Cellular: Reset operator properties on SIM change
Currently it is possible for home_provider_info and
serving_operator_info to contain stale properties.

This can be a problem when the active SIM changes to a
locked SIM with only partial information, causing stale
properties from the previous active SIM to be applied.

This resets the operator info when the primary SIM changes,
before applying the operator properties.

It also fixes an issue where the MobileOperatorInfo
instances were not being correctly observed after a Modem
restart (e.g. after a slot change or Inhibit operation).

BUG=b:187759962
TEST=tast tests and manual testing

Change-Id: I6507ac9cd61fb8a45ed3c762353162913574d0c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2895923
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>