debugd: Add bt_quality_report to FeedbackLogs
When a user files a feedback report, chrome will send a dbus
method, GetBigFeedbackLogs, to debugd. The debugd will in turn
call CreateBluetoothReport to create Bluetooth related logs.

This patch adds a new bluetooth_utils.cc to support the function
to generate a standalone file, bluetooth_quality_report, for the
feedback report.

Note that the bluetooth_quality_report file will be generated
only when the user logins with a Googler account. For a
non-Googler, the directory /var/log/bluetooth does not exist.
And hence, it will just return without generating any Bluetooth
report.

BUG=b:188508796
TEST=Perform the steps below.
Emerge, deploy, and restart debugd.

On the DUT, press shift-alt-i to file a feedback.
- If the user is a Googler, confirm the existence of
  /var/log/bluetooth/bt_quality_report.
- For a non-Googler, confirm that the /var/log/bluetooth
  directory does not exist.

Cq-Depend: chromium:2909739, chromium:2921688
Change-Id: Ie4776854d7245469116900ac21987d6af997d2e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2902550
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>