vm_tools: Add a launch command to VMC
See go/vmc-launch for details.

The 'launch' command is a product-focused VMC utility that is designed
to run VM products (like crostini). It wraps several other functions in
VMC, in order to launch the VM in a way that most-closely resembles how
chrome would launch it.

It is similar to 'start', in the sense that both initialize VMs, but
launch is separated from start because:
 - it wraps multiple VMC commands, not just start (e.g. in the case of
   'launch crostini', both 'start' and 'container' are called)
 - it focuses on products rather than features/characteristics of VMs.

BUG=b:186587105, b:177367054
TEST=vmc launch borealis, crostini, and termina

Change-Id: I12679ac32d3472c099176d6c3937e2f65e0a9c64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2898407
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>