camera: add cc_embed_data GN template
The `cc_embed_data` GN template can be used to generate C++ header and
source files that embed a list of text files.  The content of the text
files can be accessed as uint8_t arrays with the generated getter
function that returns a EmbeddedFileToc instance.

BUG=b:187157014
TEST=`FEATURES=test emerge-volteer cros-camera-libs`

Change-Id: I53a88b638512903e0062cab0f9ac9ada5de6610b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2902546
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>