hps: Add module lock and refactor feature enable
Add a lock for methods when accessing the HPS hardware,
and refactor the feature control to use the lock to enable/disable features.
This is to ensure that only one thread is accessing the hardware at a
time.

Ensure that all public methods use the lock, which
requires wrapping the internal download method so that
it can be accessed from the booting code.

BUG=b:186610248
TEST=Build and run test utils

Change-Id: I55adffa503f452a654bc51006dbb66aeabf52e8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2903611
Reviewed-by: Alan Green <avg@chromium.org>
Tested-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Andrew McRae <amcrae@google.com>