camera: add a workaround for multiple device open
It seems that for some reason the request or result handler thread of
the HAL, and as a result the Mojo channel can be blocked which disallows
the camera client to properly close the camera under use.

There isn't really anything "right" we can do about this. In theory
multiple opens could be legit as two different apps / VMs may try to
access the same camera concurrently. If the HAL thread is wedged, the
only way to recover from that is to restart the whole cros-camera
process, unless we can move cros-camera to a multi-process architecture.

As a stopgap, we check force restart the cros-camera service if the
following conditions are met:

- the HAL adapter gets multiple open calls to the same device, and
- the request of result monitors of that device aren't kicked for quite
  some time

BUG=b:155830039
TEST=Hack cros-camera to block the capture request thread, and make
sure the unclosed device can be re-opened by Chrome.

Change-Id: I83d17e78c9401082c42985cfd1cbf4b4231b82a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2903623
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>