ml: Allow Handwriting Service to accepts path.
This change expands the ML Service API for Handwriting. It allows the
client to pass the path on disk where the language models are store.

This patch is part of the Language Packs project.
In this change we add a new parameter to the Mojo call to ML Service
that contains the path of the model files. We plan to deprecate the
previous version of LoadHandwritingModel API that reads everything from
rootfs. All clients of Handwriting will use Language Packs and the new
API.

All changes are behind a false flag, so it's a no-op.

BUG=b:186469544
TEST=precq passes, deployed on DUT to verify that existing functionality
still works

Change-Id: Id7d2f3133f99e1ed94c152c1ae773503c265aca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2903625
Commit-Queue: Claudio M <claudiomagni@chromium.org>
Tested-by: Claudio M <claudiomagni@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>