Add DBus signal for debugd's packet capture
Add a signal to system_api's DBus API and emit the DBus signal when the
packet capture operation starts or stops in debugd. The signal will be
received from Chrome to show a notification to indicate the user that
there's an ongoing packet capture operation.

BUG=b:181525070
TEST=None.

Change-Id: I402fe4c5b06e293c3085de3a8c4881e1d2d7e4f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2912099
Reviewed-by: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Irem Uğuz <iremuguz@google.com>
Commit-Queue: Irem Uğuz <iremuguz@google.com>