ml: Add field enable_formatting to soda config, and make it effective.
Adds a field, 'enable_formatting' for automated formatting to the soda
configuration mojo. This becomes the field of the same name in the
underlying protocol buffer.

In order to maintain stability, adds a new enum "OptionalBool" which
is used in place of a native bool, and contains a 3rd type:
"unset". By default, we consider true the default for prod purposes.

BUG=chromium:1212911
TEST=All unit tests pass.

Change-Id: I6f040f625089f08ea5affb5a24ff5e2a60970bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2914938
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Rob Schonberger <robsc@chromium.org>
Auto-Submit: Rob Schonberger <robsc@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>